Changed localMemorySize field from Int to Long.
git-svn-id: https://kask.eti.pg.gda.pl/svn/kernelhive/KernelHive/src@270 1473d755-6451-4e3e-83a9-0691c74848a3

+++ hive-jcommon/trunk/src/pl/gda/pg/eti/kernelhive/common/clusterService/Device.java
@@ -11,7 +11,7 @@ public class Device {
 	public int computeUnitsNumber;
 	public int clock;
 	public Long globalMemoryBytes;
-	public int localMemoryBytes;
+	public Long localMemoryBytes;
 	public int workGroupSize;
 	
 	public boolean busy = false;
@@ -37,7 +37,7 @@ public class Device {
 		computeUnitsNumber = Integer.parseInt(parameters[3]);
 		clock = Integer.parseInt(parameters[4]);
 		globalMemoryBytes = Long.parseLong(parameters[5]);
-		localMemoryBytes = Integer.parseInt(parameters[6]);
+		localMemoryBytes = Long.parseLong(parameters[6]);
 		workGroupSize = Integer.parseInt(parameters[7]);
 	}
